API_USUARIO/requirements.txt
mysql-connector-python
fastapi 
uvicorn 
# "fastapi[standard]"
requests 
pydantic-settings 